First Commit. Refactored AbstractEntity away. Boots now.
