--- src/main/java/com/example/helpingout/models/AbstractEntity.java
@@ -1,31 +1,0 @@
-package com.example.helpingout.models;
-
-import javax.persistence.GeneratedValue;
-import javax.persistence.Id;
-import javax.persistence.MappedSuperclass;
-import java.util.Objects;
-
-@MappedSuperclass
-public abstract class AbstractEntity {
-
-    @Id
-    @GeneratedValue
-    private int id;
-
-    public int getId() {
-        return id;
-    }
-
-    @Override
-    public boolean equals(Object o) {
-        if (this == o) return true;
-        if (o == null || getClass() != o.getClass()) return false;
-        AbstractEntity that = (AbstractEntity) o;
-        return id == that.id;
-    }
-
-    @Override
-    public int hashCode() {
-        return Objects.hash(id);
-    }
-}
