+++ src/main/resources/application.properties
@@ -1,7 +1,7 @@
 # Database connection settings
 spring.datasource.url=jdbc:mysql://localhost:3306/helping-out
-spring.datasource.username=helpingout
-spring.datasource.password=helpingout
+spring.datasource.username=helping-out
+spring.datasource.password=password
 
 # Specify the DBMS
 spring.jpa.database = MYSQL
